Issue 16: Placement of fixed-size side-by-side inserts fails unless the first in a row is the shortest
Correct the height of a layout "section", so that subsequent rows are placed in the correct vertical position on the page, to maintain the relative layout of the fixed-sized inserts.

This avoids the case where articles are placed overlapping after the FixedElementsRelativeLayout has already checked that they will fit on the page in the correct relative position.