Issue 17: Wiggly bottom lines
Reworked how articles are fitted into their columns.

A new setting "maxSquashVSpace" determines how close an article end needs to be to the next element in order to "snap" to the column.

A new macro, \stretchsplitbox, will set an article fragment to a certain prescribed height, used if this "snap" applies.
This uses the \vbox to \len{\unvbox\id} trick to attempt to stretch the \vbox to the correct height. This doesn't always work; sometimes the unfrozen glue is stretched to a greater height than was asked for and an overfull vbox is produced, but it seems to work very well for normal article text.
We keep \copy instead of \unvbox for the normal case, as this is more reliable for producing output the same length as the length calculator predicted.

It's not perfect, but most of the time it keeps the bottom edge straight, which is what you want for a newspaper.