Issue 18: Error placing article fragment when a column is almost completely filled
Updated the first-fit algorithm in this case.
After placing the alley, this will no longer split the article and simply proceed with the next column or page.

In the case where the alley is a line, this will work well: a visual line placed at the end of the column to take up space.
In the case where the alley is space, this will leave a gap at the end of the column. There's no good solution to this other than lengthening the previous item. So in this case, spit out a warning.