Issue 20: Missing splitable content when page is almost filled
If an overflow is generated, stop trying to fit content onto the page, and return anything unprocessed to the queue.
This way, we avoid data loss.
The first fit algorithm should only produce an overflow when the page is full (baring large unbreakable elements like headings), so the chance of being able to fit any of the next article is vanishingly small regardless.